+++ skrub/_data_ops/_skrub_namespace.py
@@ -147,6 +147,7 @@ class SkrubNamespace:
         estimator,
         y=None,
         cols=_SELECT_ALL_COLUMNS,
+        no_wrap=False,
         how="auto",
         allow_reject=False,
         unsupervised=False,
@@ -160,6 +161,7 @@ class SkrubNamespace:
                 cols=cols,
                 X=self._data_op,
                 y=y,
+                no_wrap=no_wrap,
                 how=how,
                 allow_reject=allow_reject,
                 unsupervised=unsupervised,
@@ -176,6 +178,7 @@ class SkrubNamespace:
         y=None,
         cols=_SELECT_ALL_COLUMNS,
         exclude_cols=None,
+        no_wrap=False,
         how="auto",
         allow_reject=False,
         unsupervised=False,
@@ -206,7 +209,20 @@ class SkrubNamespace:
             _not_ be applied. The columns that are matched by ``cols`` AND not
             matched by ``exclude_cols`` are transformed.
 
+        no_wrap : bool, default = False
+            Disable wrapping of transformers in :class:`ApplyToCols`.
+
+            By default, when ``estimator`` is a transformer and the input is a
+            DataFrame, the transformer is wrapped in an instance of
+            :class:`ApplyToCols`, which allows applying it to part of the
+            dataframe only through the ``cols`` and ``allow_reject``
+            parameters. Passing ``no_wrap=True`` disables this wrapping in all
+            cases. When ``no_wrap`` is True, ``cols`` and ``allow_reject``
+            cannot be used.
+
         how : "auto", "cols", "frame" or "no_wrap", optional
+            Deprecated. Use ``no_wrap`` instead.
+
             How the estimator is applied. In most cases the default "auto"
             is appropriate.
 
@@ -225,6 +241,8 @@ class SkrubNamespace:
               transformer, the "no_wrap" strategy is chosen. Otherwise if the
               estimator has a ``__single_column_transformer__`` attribute,
               "cols" is chosen. Otherwise "frame" is chosen.
+
+            .. deprecated:: 0.9.0
 
         allow_reject : bool, optional
             Whether the transformer can refuse to transform columns for which
@@ -439,6 +457,7 @@ class SkrubNamespace:
             estimator=estimator,
             y=y,
             cols=cols,
+            no_wrap=no_wrap,
             how=how,
             allow_reject=allow_reject,
             unsupervised=unsupervised,
@@ -694,7 +713,7 @@ class SkrubNamespace:
         2     cup  2020-04-04
         3   spoon  2020-04-05
         """
-        return self._apply(SelectCols(cols), how="no_wrap")
+        return self._apply(SelectCols(cols), no_wrap=True)
 
     @checked_data_op_constructor
     def drop(self, cols):
@@ -747,7 +766,7 @@ class SkrubNamespace:
         2   3         5
         3   4         1
         """
-        return self._apply(DropCols(cols), how="no_wrap")
+        return self._apply(DropCols(cols), no_wrap=True)
 
     @checked_data_op_constructor
     def concat(self, others, axis=0):
@@ -3203,15 +3222,13 @@ class SkrubNamespace:
         <AppliedEstimator>
         Result:
         ―――――――
-        ApplyToSubFrame(transformer=TableVectorizer())
+        ApplyToCols(transformer=TableVectorizer())
 
         Note that in order to restrict transformers to a subset of columns,
-        they will be wrapped in a meta-estimator ``ApplyToSubFrame`` or
-        ``ApplyToEachCol`` depending if the transformer is applied to each column
-        separately or not. The actual transformer can be retrieved through the
-        ``transformer_`` attribute of ``ApplyToSubFrame`` or ``transformers_``
-        attribute of ``ApplyToEachCol`` (a dictionary mapping column names to the
-        corresponding transformer).
+        they will be wrapped in a meta-estimator :class:`ApplyToCols`. The
+        actual transformer can be retrieved through the ``transformer_`` (or
+        ``transformers_``, in the case of single-column transformers); see the
+        documentation of :class:`ApplyToCols` for details.
 
         >>> fitted_vectorizer.transformer_
         <GetAttr 'transformer_'>
@@ -3237,8 +3254,8 @@ class SkrubNamespace:
         <AppliedEstimator>
         Result:
         ―――――――
-        ApplyToEachCol(cols=(string() - cols('date')),
-                     transformer=StringEncoder(n_components=2))
+        ApplyToCols(cols=(string() - cols('date')),
+                    transformer=StringEncoder(n_components=2))
         >>> fitted_vectorizer.transformers_
         <GetAttr 'transformers_'>
         Result: